requirements.txt
flask==3.0.0
Flask-Cors==4.0.0
ffmpeg-python==0.2.0
memory-profiler==0.61.0
numpy==1.24.4
openai-whisper==20231117
regex==2023.10.3
waitress==2.1.2
werkzeug==3.0.1